src/requirements.txt
streamlit==1.32.0
pandas==2.2.0
numpy==1.26.4
scikit-learn==1.4.0
transformers==4.38.0
torch==2.2.0
groq==0.4.2
psycopg2-binary==2.9.9
sqlalchemy==2.0.28
python-dotenv==1.0.1
plotly==5.20.0
jira==3.6.0
pydantic==2.6.1
loguru==0.7.2
pytest==8.0.2
httpx==0.27.0